Try to resolve possible circular import
